refactor(orga): merge campaign creation new route and new-catalogue route
