+++ orga/app/controllers/authenticated/campaigns/new.js
@@ -8,10 +8,16 @@ export default class NewController extends Controller {
   @service store;
   @service notifications;
   @service intl;
+  @service featureToggles;
 
   @tracked errors;
+  @tracked courseId = null;
 
-  queryParams = ['source'];
+  queryParams = ['source', 'courseId'];
+
+  get catalogueFeatureEnabled() {
+    return Boolean(this.featureToggles.featureToggles?.displayCatalogue);
+  }
 
   @action
   async createCampaign() {
