+++ orga/tests/acceptance/campaign-creation-test.js
@@ -16,171 +16,425 @@ module('Acceptance | Campaign Creation', function (hooks) {
   setupApplicationTest(hooks);
   setupMirage(hooks);
   setupIntl(hooks);
+  module('when feature toggle displayCatalogue is false', function () {
+    test('it should not be accessible by an unauthenticated user', async function (assert) {
+      // when
+      await visit('/campagnes/creation');
 
-  test('it should not be accessible by an unauthenticated user', async function (assert) {
-    // when
-    await visit('/campagnes/creation');
+      // then
+      assert.strictEqual(currentURL(), '/connexion');
+    });
 
-    // then
-    assert.strictEqual(currentURL(), '/connexion');
-  });
-
-  test('it should allow to create a combined course and redirect to the newly created combined course', async function (assert) {
-    // given
-    const user = createUserWithMembershipAndTermsOfServiceAccepted();
-    server.create('member-identity', { id: user.id, firstName: user.firstName, lastName: user.lastName });
-    createPrescriberByUser({ user });
-
-    await authenticateSession(user.id);
-    availableCombinedCourseBlueprints = server.createList('combined-course-blueprint', 2);
-    const expectedCombinedCourseBlueprintName = availableCombinedCourseBlueprints[1].name;
-
-    const screen = await visit('/campagnes/creation');
-    await fillByLabel('Nom de la campagne *', 'Mon parcours combiné');
-    await clickByName(t('pages.campaign-creation.purpose.combined-course'));
-
-    await click(screen.getByLabelText(`${t('pages.campaign-creation.combined-course-blueprints-list-label')} *`));
-    await click(await screen.findByRole('option', { description: expectedCombinedCourseBlueprintName }));
-
-    // when
-    await clickByName('Créer la campagne');
-    // then
-    assert.strictEqual(server.db.campaigns[0].name, 'Mon parcours combiné');
-    assert.strictEqual(currentURL(), `/parcours/${server.db.campaigns[0].id}`);
-  });
-
-  module('when the prescriber is authenticated', (hooks) => {
-    hooks.beforeEach(async () => {
-      availableTargetProfiles = server.createList('target-profile', 2);
-
+    test('it should allow to create a combined course and redirect to the newly created combined course', async function (assert) {
+      // given
       const user = createUserWithMembershipAndTermsOfServiceAccepted();
       server.create('member-identity', { id: user.id, firstName: user.firstName, lastName: user.lastName });
       createPrescriberByUser({ user });
 
       await authenticateSession(user.id);
-    });
+      availableCombinedCourseBlueprints = server.createList('combined-course-blueprint', 2);
+      const expectedCombinedCourseBlueprintName = availableCombinedCourseBlueprints[1].name;
 
-    hooks.afterEach(function () {
-      const notificationMessagesService = this.owner.lookup('service:notifications');
-      notificationMessagesService.clearAll();
-    });
-
-    test('it should be accessible for an authenticated prescriber', async function (assert) {
-      // when
       const screen = await visit('/campagnes/creation');
+
+      await fillByLabel('Nom de la campagne *', 'Mon parcours combiné');
+      await clickByName(t('pages.campaign-creation.purpose.combined-course'));
+
+      await click(screen.getByLabelText(`${t('pages.campaign-creation.combined-course-blueprints-list-label')} *`));
+      await click(await screen.findByRole('option', { description: expectedCombinedCourseBlueprintName }));
+
+      // when
+      await clickByName('Créer la campagne');
+      // then
+      assert.strictEqual(server.db.campaigns[0].name, 'Mon parcours combiné');
+      assert.strictEqual(currentURL(), `/parcours/${server.db.campaigns[0].id}`);
+    });
+
+    module('when the prescriber is authenticated', (hooks) => {
+      hooks.beforeEach(async () => {
+        availableTargetProfiles = server.createList('target-profile', 2);
+
+        const user = createUserWithMembershipAndTermsOfServiceAccepted();
+        server.create('member-identity', { id: user.id, firstName: user.firstName, lastName: user.lastName });
+        createPrescriberByUser({ user });
+
+        await authenticateSession(user.id);
+      });
+
+      hooks.afterEach(function () {
+        const notificationMessagesService = this.owner.lookup('service:notifications');
+        notificationMessagesService.clearAll();
+      });
+
+      test('it should be accessible for an authenticated prescriber', async function (assert) {
+        // when
+        const screen = await visit('/campagnes/creation');
+
+        // then
+        assert.strictEqual(currentURL(), '/campagnes/creation');
+        assert.ok(screen.getByText("Création d'une campagne"));
+      });
+
+      test('it should allow to create a campaign of type ASSESSMENT and redirect to the newly created campaign', async function (assert) {
+        // given
+        const expectedTargetProfileId = availableTargetProfiles[1].id;
+        const expectedTargetProfileName = availableTargetProfiles[1].name;
+
+        const screen = await visit('/campagnes/creation');
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+        await clickByName('Évaluer les participants');
+        await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
+        await click(await screen.findByRole('option', { description: expectedTargetProfileName }));
+
+        const externalIdentifier = screen
+          .getByText(t('pages.campaign-creation.external-id-label.question-label'))
+          .closest('fieldset');
+        const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
+        await click(element);
+
+        // when
+        await clickByName('Créer la campagne');
+
+        // then
+        const firstCampaign = server.db.campaigns[0];
+        assert.strictEqual(firstCampaign.name, 'Ma Campagne');
+        assert.strictEqual(firstCampaign.targetProfileId, expectedTargetProfileId);
+        assert.strictEqual(currentURL(), '/campagnes/1/parametres');
+      });
+
+      test('it should allow to create a campaign of type PROFILES_COLLECTION and redirect to the newly created campaign', async function (assert) {
+        // given
+        const screen = await visit('/campagnes/creation');
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+        await clickByName('Collecter les profils Pix des participants');
+        const externalIdentifier = screen
+          .getByText(t('pages.campaign-creation.external-id-label.question-label'))
+          .closest('fieldset');
+        const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
+        await click(element);
+
+        // when
+        await clickByName('Créer la campagne');
+
+        // then
+        assert.strictEqual(server.db.campaigns[0].name, 'Ma Campagne');
+        assert.strictEqual(currentURL(), '/campagnes/1/parametres');
+      });
+
+      test('it should create campaign if user changes type after filling the form', async function (assert) {
+        // given
+        const expectedTargetProfileName = availableTargetProfiles[1].name;
+
+        const screen = await visit('/campagnes/creation');
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+        await clickByName('Évaluer les participants');
+        await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
+        await click(await screen.findByRole('option', { description: expectedTargetProfileName }));
+
+        const title = `${t('pages.campaign-creation.course-title.label')} ${t('pages.campaign-creation.course-title.sublabel')}`;
+
+        await fillByLabel(title, 'Savoir rechercher');
+        await clickByName('Non');
+
+        // when
+        await clickByName(t('pages.campaign-creation.actions.create'));
+
+        // then
+        assert.strictEqual(server.db.campaigns[0].name, 'Ma Campagne');
+        assert.strictEqual(currentURL(), '/campagnes/1/parametres');
+      });
+
+      test('it should set the current user as owner by default when creating a campaign', async function (assert) {
+        // given
+        const targetProfileName = availableTargetProfiles[1].name;
+        const screen = await visit('/campagnes/creation');
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+        await clickByName('Évaluer les participants');
+        await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
+        await click(await screen.findByRole('option', { description: targetProfileName }));
+
+        // when
+        await clickByName('Créer la campagne');
+
+        // then
+        assert.ok(screen.getByText('Harry Cover', { selector: 'dd' }));
+      });
+
+      test('it should display error on global form when error 500 is returned from backend', async function (assert) {
+        // given
+        const screen = await visit('/campagnes/creation');
+
+        const expectedTargetProfileName = availableTargetProfiles[1].name;
+        server.post('/campaigns', {}, 500);
+
+        // when
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+        await clickByName('Évaluer les participants');
+        await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
+        await click(await screen.findByRole('option', { description: expectedTargetProfileName }));
+        const externalIdentifier = screen
+          .getByText(t('pages.campaign-creation.external-id-label.question-label'))
+          .closest('fieldset');
+        const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
+        await click(element);
+        await clickByName('Créer la campagne');
+
+        // then
+        assert.strictEqual(currentURL(), '/campagnes/creation');
+        assert.ok(screen.getByText('Une erreur est survenue. Veuillez réessayer ultérieurement.'));
+      });
+    });
+  });
+  module('when feature toggle displayCatalogue is true', function (hooks) {
+    hooks.beforeEach(function () {
+      server.create('feature-toggle', { id: '0', displayCatalogue: true });
+
+      availableCombinedCourseBlueprints = server.createList('course', 2, { type: 'blueprint' });
+      availableCombinedCourseBlueprints.forEach((course) => {
+        server.create('combined-course-blueprint-overview', { id: course.id, name: course.name });
+        server.create('combined-course-blueprint', { id: course.id, name: course.name });
+      });
+
+      availableTargetProfiles = server.createList('course', 2, { type: 'targetProfile' });
+      availableTargetProfiles.forEach((course) => {
+        server.create('target-profile-overview', { id: course.id, name: course.name });
+        server.create('target-profile', { id: course.id, name: course.name });
+      });
+    });
+
+    test('it should not be accessible by an unauthenticated user', async function (assert) {
+      // when
+      await visit('/campagnes/creation');
 
       // then
-      assert.strictEqual(currentURL(), '/campagnes/creation');
-      assert.ok(screen.getByText("Création d'une campagne"));
+      assert.strictEqual(currentURL(), '/connexion');
     });
 
-    test('it should allow to create a campaign of type ASSESSMENT and redirect to the newly created campaign', async function (assert) {
+    test('it should allow to create a combined course and redirect to the newly created combined course', async function (assert) {
       // given
-      const expectedTargetProfileId = availableTargetProfiles[1].id;
-      const expectedTargetProfileName = availableTargetProfiles[1].name;
+      const user = createUserWithMembershipAndTermsOfServiceAccepted();
+      server.create('member-identity', { id: user.id, firstName: user.firstName, lastName: user.lastName });
+      createPrescriberByUser({ user });
+
+      await authenticateSession(user.id);
+
+      const expectedCombinedCourseBlueprintName = availableCombinedCourseBlueprints[1].name;
 
       const screen = await visit('/campagnes/creation');
-      await fillByLabel('Nom de la campagne *', 'Ma Campagne');
-      await clickByName('Évaluer les participants');
-      await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
-      await click(await screen.findByRole('option', { description: expectedTargetProfileName }));
 
-      const externalIdentifier = screen
-        .getByText(t('pages.campaign-creation.external-id-label.question-label'))
-        .closest('fieldset');
-      const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
-      await click(element);
+      await click(screen.getByRole('radio', { name: t('pages.campaign-creation.purpose.combined-course') }));
+
+      await click(
+        screen.getByRole('link', {
+          name: t('pages.campaign-creation.course-selection-label'),
+        }),
+      );
+
+      await click(
+        screen.getByRole('link', {
+          name: t('pages.catalogue.modal.open-modal', { name: expectedCombinedCourseBlueprintName }),
+        }),
+      );
+
+      const dialog = await screen.findByRole('dialog');
+      await click(within(dialog).getByRole('link', { name: t('pages.catalogue.modal.select-course') }));
+
+      await fillByLabel('Nom de la campagne *', 'Mon parcours combiné');
 
       // when
       await clickByName('Créer la campagne');
 
       // then
-      const firstCampaign = server.db.campaigns[0];
-      assert.strictEqual(firstCampaign.name, 'Ma Campagne');
-      assert.strictEqual(firstCampaign.targetProfileId, expectedTargetProfileId);
-      assert.strictEqual(currentURL(), '/campagnes/1/parametres');
+      assert.strictEqual(server.db.campaigns[0].name, 'Mon parcours combiné');
+      assert.strictEqual(currentURL(), `/parcours/${server.db.campaigns[0].id}`);
     });
 
-    test('it should allow to create a campaign of type PROFILES_COLLECTION and redirect to the newly created campaign', async function (assert) {
-      // given
-      const screen = await visit('/campagnes/creation');
-      await fillByLabel('Nom de la campagne *', 'Ma Campagne');
-      await clickByName('Collecter les profils Pix des participants');
-      const externalIdentifier = screen
-        .getByText(t('pages.campaign-creation.external-id-label.question-label'))
-        .closest('fieldset');
-      const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
-      await click(element);
+    module('when the prescriber is authenticated', (hooks) => {
+      hooks.beforeEach(async () => {
+        const user = createUserWithMembershipAndTermsOfServiceAccepted();
+        server.create('member-identity', { id: user.id, firstName: user.firstName, lastName: user.lastName });
+        createPrescriberByUser({ user });
 
-      // when
-      await clickByName('Créer la campagne');
+        await authenticateSession(user.id);
+      });
 
-      // then
-      assert.strictEqual(server.db.campaigns[0].name, 'Ma Campagne');
-      assert.strictEqual(currentURL(), '/campagnes/1/parametres');
-    });
+      hooks.afterEach(function () {
+        const notificationMessagesService = this.owner.lookup('service:notifications');
+        notificationMessagesService.clearAll();
+      });
 
-    test('it should create campaign if user changes type after filling the form', async function (assert) {
-      // given
-      const expectedTargetProfileName = availableTargetProfiles[1].name;
+      test('it should be accessible for an authenticated prescriber', async function (assert) {
+        // when
+        const screen = await visit('/campagnes/creation');
 
-      const screen = await visit('/campagnes/creation');
-      await fillByLabel('Nom de la campagne *', 'Ma Campagne');
-      await clickByName('Évaluer les participants');
-      await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
-      await click(await screen.findByRole('option', { description: expectedTargetProfileName }));
+        // then
+        assert.strictEqual(currentURL(), '/campagnes/creation');
+        assert.ok(screen.getByText("Création d'une campagne"));
+      });
 
-      const title = `${t('pages.campaign-creation.course-title.label')} ${t('pages.campaign-creation.course-title.sublabel')}`;
+      test('it should allow to create a campaign of type ASSESSMENT and redirect to the newly created campaign', async function (assert) {
+        // given
+        const expectedTargetProfileId = availableTargetProfiles[1].id;
+        const expectedTargetProfileName = availableTargetProfiles[1].name;
 
-      await fillByLabel(title, 'Savoir rechercher');
-      await clickByName('Non');
+        const screen = await visit('/campagnes/creation');
 
-      // when
-      await clickByName(t('pages.campaign-creation.actions.create'));
+        await click(screen.getByRole('radio', { name: t('pages.campaign-creation.purpose.assessment') }));
 
-      // then
-      assert.strictEqual(server.db.campaigns[0].name, 'Ma Campagne');
-      assert.strictEqual(currentURL(), '/campagnes/1/parametres');
-    });
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.campaign-creation.course-selection-label'),
+          }),
+        );
 
-    test('it should set the current user as owner by default when creating a campaign', async function (assert) {
-      // given
-      const targetProfileName = availableTargetProfiles[1].name;
-      const screen = await visit('/campagnes/creation');
-      await fillByLabel('Nom de la campagne *', 'Ma Campagne');
-      await clickByName('Évaluer les participants');
-      await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
-      await click(await screen.findByRole('option', { description: targetProfileName }));
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.catalogue.modal.open-modal', { name: expectedTargetProfileName }),
+          }),
+        );
+        const dialog = await screen.findByRole('dialog');
+        await click(within(dialog).getByRole('link', { name: t('pages.catalogue.modal.select-course') }));
 
-      // when
-      await clickByName('Créer la campagne');
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
 
-      // then
-      assert.ok(screen.getByText('Harry Cover', { selector: 'dd' }));
-    });
+        const externalIdentifier = screen
+          .getByText(t('pages.campaign-creation.external-id-label.question-label'))
+          .closest('fieldset');
+        const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
+        await click(element);
 
-    test('it should display error on global form when error 500 is returned from backend', async function (assert) {
-      // given
-      const screen = await visit('/campagnes/creation');
+        // when
+        await clickByName('Créer la campagne');
 
-      const expectedTargetProfileName = availableTargetProfiles[1].name;
-      server.post('/campaigns', {}, 500);
+        // then
+        const firstCampaign = server.db.campaigns[0];
+        assert.strictEqual(firstCampaign.name, 'Ma Campagne');
+        assert.strictEqual(firstCampaign.targetProfileId, expectedTargetProfileId);
+        assert.strictEqual(currentURL(), '/campagnes/1/parametres');
+      });
 
-      // when
-      await fillByLabel('Nom de la campagne *', 'Ma Campagne');
-      await clickByName('Évaluer les participants');
-      await click(screen.getByLabelText(`${t('pages.campaign-creation.target-profiles-list-label')} *`));
-      await click(await screen.findByRole('option', { description: expectedTargetProfileName }));
-      const externalIdentifier = screen
-        .getByText(t('pages.campaign-creation.external-id-label.question-label'))
-        .closest('fieldset');
-      const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
-      await click(element);
-      await clickByName('Créer la campagne');
+      test('it should allow to create a campaign of type PROFILES_COLLECTION and redirect to the newly created campaign', async function (assert) {
+        // given
+        const screen = await visit('/campagnes/creation');
+        await clickByName('Collecter les profils Pix des participants');
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+        const externalIdentifier = screen
+          .getByText(t('pages.campaign-creation.external-id-label.question-label'))
+          .closest('fieldset');
+        const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
+        await click(element);
 
-      // then
-      assert.strictEqual(currentURL(), '/campagnes/creation');
-      assert.ok(screen.getByText('Une erreur est survenue. Veuillez réessayer ultérieurement.'));
+        // when
+        await clickByName('Créer la campagne');
+
+        // then
+        assert.strictEqual(server.db.campaigns[0].name, 'Ma Campagne');
+        assert.strictEqual(currentURL(), '/campagnes/1/parametres');
+      });
+
+      test('it should create campaign if user changes type after filling the form', async function (assert) {
+        // given
+        const expectedTargetProfileName = availableTargetProfiles[1].name;
+
+        const screen = await visit('/campagnes/creation');
+
+        await click(screen.getByRole('radio', { name: t('pages.campaign-creation.purpose.assessment') }));
+
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.campaign-creation.course-selection-label'),
+          }),
+        );
+
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.catalogue.modal.open-modal', { name: expectedTargetProfileName }),
+          }),
+        );
+        const dialog = await screen.findByRole('dialog');
+        await click(within(dialog).getByRole('link', { name: t('pages.catalogue.modal.select-course') }));
+
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+
+        const title = `${t('pages.campaign-creation.course-title.label')} ${t('pages.campaign-creation.course-title.sublabel')}`;
+
+        await fillByLabel(title, 'Savoir rechercher');
+        await clickByName('Non');
+
+        // when
+        await clickByName(t('pages.campaign-creation.actions.create'));
+
+        // then
+        assert.strictEqual(server.db.campaigns[0].name, 'Ma Campagne');
+        assert.strictEqual(currentURL(), '/campagnes/1/parametres');
+      });
+
+      test('it should set the current user as owner by default when creating a campaign', async function (assert) {
+        // given
+        const targetProfileName = availableTargetProfiles[1].name;
+        const screen = await visit('/campagnes/creation');
+
+        await click(screen.getByRole('radio', { name: t('pages.campaign-creation.purpose.assessment') }));
+
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.campaign-creation.course-selection-label'),
+          }),
+        );
+
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.catalogue.modal.open-modal', { name: targetProfileName }),
+          }),
+        );
+        const dialog = await screen.findByRole('dialog');
+        await click(within(dialog).getByRole('link', { name: t('pages.catalogue.modal.select-course') }));
+
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+
+        // when
+        await clickByName('Créer la campagne');
+
+        // then
+        assert.ok(screen.getByText('Harry Cover', { selector: 'dd' }));
+      });
+
+      test('it should display error on global form when error 500 is returned from backend', async function (assert) {
+        // given
+        const screen = await visit('/campagnes/creation');
+
+        const expectedTargetProfileName = availableTargetProfiles[1].name;
+        server.post('/campaigns', {}, 500);
+
+        await click(screen.getByRole('radio', { name: t('pages.campaign-creation.purpose.assessment') }));
+
+        // when
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.campaign-creation.course-selection-label'),
+          }),
+        );
+
+        await click(
+          screen.getByRole('link', {
+            name: t('pages.catalogue.modal.open-modal', { name: expectedTargetProfileName }),
+          }),
+        );
+        const dialog = await screen.findByRole('dialog');
+        await click(within(dialog).getByRole('link', { name: t('pages.catalogue.modal.select-course') }));
+
+        await fillByLabel('Nom de la campagne *', 'Ma Campagne');
+        const externalIdentifier = screen
+          .getByText(t('pages.campaign-creation.external-id-label.question-label'))
+          .closest('fieldset');
+        const element = within(externalIdentifier).getByRole('radio', { name: 'Non' });
+        await click(element);
+        await clickByName('Créer la campagne');
+
+        // then
+        assert.strictEqual(currentURL(), '/campagnes/creation?courseId=4');
+        assert.ok(screen.getByText('Une erreur est survenue. Veuillez réessayer ultérieurement.'));
+      });
     });
   });
 });
